+++ client/src/pages/Mypage/Mypage.tsx
@@ -1,46 +1,43 @@
-import { userState } from "#atoms/userState";
-import Button from "#components/Button/Button";
 import Header from "#components/Header/Header";
-import useHttpGet from "#hooks/http/useHttpGet";
-import { useNavigate } from "react-router-dom";
-import { useRecoilValue, useResetRecoilState } from "recoil";
-import styled from "styled-components";
-import { flexColumnSpaceBetween } from "styles/flex";
+import { useState } from "react";
+import CourseCard from "#components/Card/CourseCard/CourseCard";
+import { MYPAGE } from "#constants/myPageOptions";
+import MyPageProfile from "./MyPageProfile";
+import { Menu, MenuWrapper } from "./MyPage.styles";
 
-const Body = styled.div`
-    ${flexColumnSpaceBetween};
-    gap: 5rem;
-    align-items: center;
-    padding: 10rem;
-`;
+const data = {
+    id: 3,
+    title: "title",
+    path: [],
+    pathLength: 0,
+    useId: "catenary",
+    hDong: {
+        name: "잠실동",
+        code: "1111111111",
+    },
+    createdAt: "createdAt",
+};
 
-const UserId = styled.div`
-    font-size: 3rem;
-`;
+const MyPage = () => {
+    const [myPageOption, setMyPageOption] = useState(MYPAGE.PROFILE);
 
-const Mypage = () => {
-    const userInfo = useRecoilValue(userState);
-    const { get } = useHttpGet();
-    const userStateReset = useResetRecoilState(userState);
-    const navigate = useNavigate();
-    const logoutOnClick = async () => {
-        try {
-            console.log(userInfo);
-            userStateReset();
-            await get("/auth/logout");
-            navigate("/");
-        } catch {}
-    };
     return (
-        <>
-            <Header text="마이페이지" />
-            <Body>
-                <UserId>{userInfo.userId}</UserId>
-                <Button width="fill" onClick={logoutOnClick}>
-                    로그아웃
-                </Button>
-            </Body>
-        </>
+        <div>
+            <Header text="마이페이지"></Header>
+            <MenuWrapper>
+                <Menu enabled={myPageOption == MYPAGE.PROFILE} onClick={() => setMyPageOption(MYPAGE.PROFILE)}>
+                    내 정보
+                </Menu>
+                <Menu enabled={myPageOption == MYPAGE.COURSES} onClick={() => setMyPageOption(MYPAGE.COURSES)}>
+                    코스 목록
+                </Menu>
+                <Menu enabled={myPageOption == MYPAGE.RECRUITS} onClick={() => setMyPageOption(MYPAGE.RECRUITS)}>
+                    참여
+                </Menu>
+            </MenuWrapper>
+            <MyPageProfile MyPageOption={myPageOption}></MyPageProfile>
+        </div>
     );
 };
-export default Mypage;
+
+export default MyPage;
